+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ## [Unreleased]
 
+### Added
+- Added handle and message for departments title 
+
 ## [2.0.5] - 2021-06-17
 ### Fixed
 - Solved problem close menu to on clicking the item